Modularize codebase
